chore: heavy refactor and bug fixing
