--- app/scaffold/ratings/list/page.tsx
@@ -1,7 +1,0 @@
-import { List } from "@/features/ratings";
-
-export const revalidate = 0
-
-export default function Page() {
-  return <List />;
-}
