--- app/scaffold/ratings/new/page.tsx
@@ -1,9 +1,0 @@
-import { Create } from "@/features/ratings";
-
-export default function New() {
-  return (
-    <div>
-      <Create />
-    </div>
-  )
-}
