--- app/scaffold/reservations/list/page.tsx
@@ -1,7 +1,0 @@
-import { List } from "@/features/reservations";
-
-export const revalidate = 0
-
-export default function Page() {
-  return <List />;
-}
