--- app/scaffold/reservations/new/page.tsx
@@ -1,9 +1,0 @@
-import { Create } from "@/features/reservations";
-
-export default function New() {
-  return (
-    <div>
-      <Create />
-    </div>
-  )
-}
