--- app/scaffold/restaurants/list/page.tsx
@@ -1,7 +1,0 @@
-import { List } from "@/features/restaurants";
-
-export const revalidate = 0
-
-export default function Page() {
-  return <List />;
-}
