--- app/scaffold/users/[id]/edit/page.tsx
@@ -1,8 +1,0 @@
-import { Edit } from "@/features/users";
-import invariant from "tiny-invariant";
-
-export default function EditPage({ params }: { params: { id: string } }) {
-  invariant(params.id, "params.id is required");
-
-  return <Edit id={params.id} />;
-}
