--- app/scaffold/users/[id]/page.tsx
@@ -1,8 +1,0 @@
-import { Detail } from "@/features/users";
-import invariant from "tiny-invariant";
-
-export default function Page({ params }: { params: { id: string } }) {
-  invariant(params.id, "params.id is required");
-
-  return <Detail id={params.id} />;
-}
