--- app/scaffold/users/list/page.tsx
@@ -1,7 +1,0 @@
-import { List } from "@/features/users";
-
-export const revalidate = 0
-
-export default function Page() {
-  return <List />;
-}
