--- app/scaffold/users/new/page.tsx
@@ -1,9 +1,0 @@
-import { Create } from "@/features/users";
-
-export default function New() {
-  return (
-    <div>
-      <Create />
-    </div>
-  )
-}
