+++ features/auth/api.tsx
@@ -1,3 +1,5 @@
+import "server-only"
+
 import { createClient } from "@/lib/supabase/admin";
 
 export function api() {
