+++ features/auth/index.tsx
@@ -1,6 +1,4 @@
 export * from './components/AuthButton'
 export * from './components/LoginForm'
 export * from './components/Account'
-export * from './utils'
 export * from './middleware'
-export * from './api'
