+++ features/auth/utils.tsx
@@ -1,4 +1,5 @@
-// server only
+import "server-only"
+
 import { redirect } from "next/navigation";
 
 import { createClient } from "@/lib/supabase/server";
