+++ features/ratings/index.tsx
@@ -1,5 +1,4 @@
 export * from "./types";
-export { api } from "./api";
 export { List } from "./components/list";
 export { Detail } from "./components/detail";
 export { Create } from "./components/create";
