+++ features/reservations/index.tsx
@@ -1,5 +1,4 @@
 export * from './types'
-export * from './api'
 export * from './components/list'
 export * from './components/detail'
 export * from './components/create'
