+++ features/reservations/server.tsx
@@ -1,0 +1,3 @@
+import "server-only"
+
+export * from './api'
