+++ features/restaurants/index.tsx
@@ -1,6 +1,2 @@
 export * from "./types";
 export { cuisineOptions } from "./constants";
-export { List } from "./components/list";
-export { Detail } from "./components/detail";
-export { RestaurantList } from "./components/RestaurantList";
-export { CuisinePill } from "./components/CuisinePill";
