+++ features/restaurants/server.tsx
@@ -1,3 +1,5 @@
+import "server-only";
+
 export { api } from "./api";
-export { Create } from "./components/create";
-export { Edit } from "./components/edit";
+export { Detail } from "./components/detail";
+export { RestaurantList } from "./components/RestaurantList";
