+++ lib/supabase/admin.ts
@@ -1,4 +1,4 @@
-"use server"
+import "server-only"
 
 // https://supabase.com/docs/reference/javascript/admin-api
 import { cookies } from 'next/headers'
@@ -6,7 +6,7 @@ import { cookies } from 'next/headers'
 import { createServerClient } from '@supabase/ssr'
 import { Database } from './database.types'
 
-export async function createClient() {
+export function createClient() {
   const cookieStore = cookies()
 
   return createServerClient<Database>(
