+++ lib/supabase/client.ts
@@ -1,3 +1,4 @@
+import "client-only"
 import { createBrowserClient } from "@supabase/ssr";
 import { Database } from "./database.types";
