+++ package.json
@@ -18,11 +18,13 @@
     "@mantine/hooks": "^7.11.2",
     "@supabase/ssr": "^0.4.0",
     "@supabase/supabase-js": "^2.44.4",
+    "client-only": "^0.0.1",
     "dayjs": "^1.11.12",
     "next": "14.2.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.2.1",
+    "server-only": "^0.0.1",
     "tiny-invariant": "^1.3.3",
     "zustand": "^4.5.4"
   },
